fix: publish logs not following env var definition
