+++ src/helper.ts
@@ -5,3 +5,13 @@ export function tryParseJson(content: string) {
     return content;
   }
 }
+
+export function merge(obj1, obj2) {
+  const merged = { ...obj1 };
+
+  for (const key in obj2) {
+    if (merged[key] === undefined || merged[key] === null)
+      merged[key] = obj2[key];
+  }
+  return merged;
+}
